regroup session tests in a single class
